% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.map}
\alias{.map}
\alias{.map_chr}
\alias{.map_lgl}
\alias{.map_int}
\alias{.map_dbl}
\title{Apply/Map Operations}
\usage{
.map(.x, .f, ...)

.map_chr(.x, .f, ...)

.map_lgl(.x, .f, ...)

.map_int(.x, .f, ...)

.map_dbl(.x, .f, ...)
}
\arguments{
\item{.x}{A (likely iterable) R object (e.g. list, atomic vector, etc.).}

\item{.f}{\code{<function>} Function to be applied to each element of \code{.x}.}

\item{...}{Optional arguments to pass to \code{.f}.}
}
\description{
Apply a function to each element of an R object.
}
\seealso{
\code{\link[=lapply]{lapply()}}, \code{\link[=vapply]{vapply()}}, \code{\link[purrr:map]{purrr::map()}}
}
