% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.map2}
\alias{.map2}
\alias{.map2_chr}
\alias{.map2_lgl}
\alias{.map2_int}
\alias{.map2_dbl}
\title{Apply/Map Operations Using 2 arguments simultaneously.}
\usage{
.map2(.x, .y, .f, ...)

.map2_chr(.x, .y, .f, ...)

.map2_lgl(.x, .y, .f, ...)

.map2_int(.x, .y, .f, ...)

.map2_dbl(.x, .y, .f, ...)
}
\arguments{
\item{.x, .y}{R objects of the same length (e.g. list, atomic vector, etc.).}

\item{.f}{\code{<function>} Function to be applied to each pair of elements in \code{.x}.}

\item{...}{Optional arguments to pass to \code{.f}.}
}
\description{
Apply a function to each pair of elements in 2 R objects.
}
