% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knappyr-package.R
\docType{package}
\name{knapplyr}
\alias{knapplyr}
\alias{knapplyr-package}
\title{knapplyr: Tools to Facilitate Robust R Programming}
\description{
An opinionated collection of functions to facilitate robust, type-safe R programming.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/knapply/knapplyr}
  \item \url{https://knapply.github.io/knapplyr}
  \item Report bugs at \url{https://github.com/knapply/knapplyr/issues}
}

}
\author{
Brendan Knapp \href{mailto:brendan.g.knapp@gmail.com}{brendan.g.knapp@gmail.com}
}
